+++ packages/containers/sandbox/package.json
@@ -35,7 +35,6 @@
     "immutable": "^3.8.1",
     "invariant": "^2.2.2",
     "lodash": "^4.17.4",
-    "memoize-one": "^4.0.0",
     "prop-types": "^15.5.10",
     "react": "^16.0.0",
     "react-ace": "5.2.0",
