+++ packages/containers/sandbox/src/app/index.js
@@ -4,10 +4,9 @@
  * and other style can override it
  */
 
+import '@talend/bootstrap-theme/src/theme/theme.scss';
 import { registerAllContainers } from '@talend/react-containers/lib/register';
 import ComponentForm from '@talend/react-containers/lib/ComponentForm';
-import '@talend/bootstrap-theme/src/theme/theme.scss';
-import './index.scss';
 import cmf from '@talend/react-cmf';
 import { createLogger } from 'redux-logger';
 import actions from './actions';
