--- packages/containers/sandbox/src/app/index.scss
@@ -1,3 +1,0 @@
-:global(.tc-header-bar) {
-    position: fixed;  // FIXME: can t find it s needed
-}
